Support IE 11. Fix error on ForgotPassword for IE and MSEdge. Refactor dev and prod webpack configurations.
