+++ .babelrc
@@ -1,5 +1,9 @@
 {
   "presets": ["env", "react"],
-  "plugins": ["transform-object-rest-spread", "transform-class-properties"]
+  "plugins": [
+    "transform-class-properties",
+    "transform-es2015-arrow-functions",
+    "transform-object-rest-spread"
+  ]
 }
