+++ app/assets/src/index.js
@@ -9,6 +9,16 @@ import "font-awesome/scss/font-awesome.scss";
 import "semantic-ui-css/semantic.min.css";
 import "./loader.scss";
 
+if (!function f() {}.name) {
+  Object.defineProperty(Function.prototype, "name", {
+    get: function() {
+      var name = (this.toString().match(/^function\s*([^\s(]+)/) || [])[1];
+      Object.defineProperty(this, "name", { value: name });
+      return name;
+    }
+  });
+}
+
 const context = require.context("./components", true, /\.(js|jsx)$/);
 const foundComponents = {};
 const contextKeys = context.keys();
@@ -22,12 +32,11 @@ contextKeys.forEach(key => {
   if (a && a.name) {
     foundComponents[a.name] = a;
   }
-}); // Turn off camelcase rule
-/* eslint camelcase: 0 */ const react_component = (
-  componentName,
-  props,
-  target
-) => {
+});
+// Turn off camelcase rule
+/* eslint camelcase: 0 */
+
+const react_component = (componentName, props, target) => {
   const matchedComponent = foundComponents[componentName];
   if (matchedComponent) {
     ReactDOM.render(
