+++ app/controllers/application_controller.rb
@@ -1,6 +1,6 @@
 class ApplicationController < ActionController::Base
   before_action :authenticate_user!
-  before_action :check_browser
+  # before_action :check_browser
   protect_from_forgery with: :exception
   before_action :check_rack_mini_profiler
