+++ app/views/devise/passwords/new.html.erb
@@ -1,7 +1,7 @@
 <div id='forgot_password_page' >
   <%= javascript_tag do %>
     react_component('ForgotPassword', {
-      errors: `<%= devise_error_messages! %>`,
+      errors: '<%= devise_error_messages! %>',
       endpoint: '<%=  password_path(resource) %>',
       emailLabel: 'user[email]',
       csrf: '<%= form_authenticity_token %>'  }, 'forgot_password_page');
